fix(app): polish Windows titlebar borders
